+++ NEWS.md
@@ -3,7 +3,7 @@
 This first release includes converters between the following data structures:
 
 * complete lists of simplices, used by the {TDA} package
-* simplex tree instances of class 'Rcpp_SimplexTree' or 'simplextree',
+* simplex tree instances of class 'Rcpp_SimplexTree',
   provided by the {simplextree} package
 * simplex trees as implemented in Python GUDHI through the {reticulate} interface
 * objects of class 'igraph', provided by the {igraph} package